+++ demisto_sdk/common/hook_validations/reputation.py
@@ -1,6 +1,6 @@
-from demisto_sdk.validation.constants import Errors
-from demisto_sdk.validation.hook_validations.base_validator import BaseValidator
-from demisto_sdk.validation.tools import print_error
+from demisto_sdk.common.constants import Errors
+from demisto_sdk.common.hook_validations.base_validator import BaseValidator
+from demisto_sdk.common.tools import print_error
 
 
 class ReputationValidator(BaseValidator):
