+++ demisto_sdk/common/schemas/canvas-context-connections.yml
