+++ demisto_sdk/common/schemas/dashboard.yml
