+++ demisto_sdk/common/schemas/integration.yml
