+++ demisto_sdk/common/schemas/layout.yml
