+++ demisto_sdk/common/schemas/playbook.yml
