+++ demisto_sdk/common/schemas/script.yml
