+++ demisto_sdk/common/schemas/widget.yml
