+++ demisto_sdk/common/scripts/update_id_set.py
@@ -6,8 +6,8 @@ from multiprocessing import Pool, cpu_count
 import time
 
 # TODO: do not import *
-from demisto_sdk.validation.constants import *
-from demisto_sdk.validation.tools import get_yaml, get_to_version, get_from_version, collect_ids,\
+from demisto_sdk.common.constants import *
+from demisto_sdk.common.tools import get_yaml, get_to_version, get_from_version, collect_ids,\
     get_script_or_integration_id, LOG_COLORS, print_color, run_command
 from demisto_sdk.yaml_tools.unifier import Unifier
