+++ demisto_sdk/core.py
@@ -11,8 +11,8 @@ import sys
 import argparse
 
 from demisto_sdk.validation.configuration import Configuration
-from demisto_sdk.validation.constants import DIR_TO_PREFIX
-from demisto_sdk.validation.tools import print_color, print_error, LOG_COLORS
+from demisto_sdk.common.constants import DIR_TO_PREFIX
+from demisto_sdk.common.tools import print_color, print_error, LOG_COLORS
 from .yaml_tools.unifier import Unifier
 from .yaml_tools.extractor import Extractor
 from .validation.file_validator import FilesValidator
