+++ tests/id_test.py
@@ -1,5 +1,5 @@
 from demisto_sdk.validation.configuration import Configuration
-from demisto_sdk.validation.hook_validations.id import IDSetValidator
+from demisto_sdk.common.hook_validations.id import IDSetValidator
 CONFIG = Configuration()
