+++ tests/structure_test.py
@@ -7,9 +7,9 @@ from typing import List, Tuple
 import pytest
 import yaml
 
-from demisto_sdk.validation.constants import DIR_LIST
-from demisto_sdk.validation.hook_validations.structure import StructureValidator
-from demisto_sdk.validation.hook_validations.tests_constants import VALID_TEST_PLAYBOOK_PATH, INVALID_PLAYBOOK_PATH, \
+from demisto_sdk.common.constants import DIR_LIST
+from demisto_sdk.common.hook_validations.structure import StructureValidator
+from demisto_sdk.common.hook_validations.tests_constants import VALID_TEST_PLAYBOOK_PATH, INVALID_PLAYBOOK_PATH, \
     VALID_INTEGRATION_TEST_PATH, VALID_INTEGRATION_ID_PATH, INVALID_INTEGRATION_ID_PATH, VALID_PLAYBOOK_ID_PATH, \
     INVALID_PLAYBOOK_ID_PATH, VALID_REPUTATION_PATH, VALID_LAYOUT_PATH, INVALID_LAYOUT_PATH, INVALID_WIDGET_PATH, \
     VALID_WIDGET_PATH, VALID_DASHBOARD_PATH, INVALID_DASHBOARD_PATH, INVALID_REPUTATION_PATH
